Early work toward actual USB device emulation
Because the ROM SIMM programmer is simply a USB CDC device, I can use
Linux's existing USB CDC gadget capability to simulate a CDC serial
port, and it will show up as an actual USB device in Linux. This is an
early crude version of logic that hooks up to this capability. So far so
good!